composable_views

Macro ui_id

ui_id!() { /* proc-macro */ }

Produces a unique id (at compile-time)

let a = ui_id::ui_id!();
let b = ui_id::ui_id!();
assert_ne!(a, b);

More complex cases are handled by passing in optional runtime values to distinguish ids that are logically different but produced by the same location in the source code. Loops, for example, can be handled by passing in the loop index.

use itertools::Itertools;
(0..1000)
    .map(|n: usize| { return ui_id::ui_id!(n) })
    .collect::<Vec<_>>()
    .into_iter()
    .combinations(2)
    .into_iter()
    .for_each(|pair| assert_ne!(pair[0], pair[1]));

The macro actually produces a NonZero u128 so that an optional id is the same size as the id itself.

let a = ui_id::ui_id!();
assert_eq!(std::mem::size_of_val(&a), size_of::<u128>());

let b = Some(ui_id::ui_id!());
assert_eq!(std::mem::size_of_val(&b), size_of::<u128>());

Since, to do so, zero values need to be protected against, ensuring that ids are also valid UUIDs is essentially free.

let ui_id = ui_id::ui_id!();
let uuid = uuid::Uuid::from_u128(ui_id.get());
assert_eq!(uuid.get_version_num(), 4);
assert_eq!(uuid.get_variant(), uuid::Variant::RFC4122);