composable_views

Module text

Structs

Family

Font

Glyphs

A buffer that contains the results of the shaping process.

Language

A script language.

Script

A text script.

Tag

A 4-byte tag.

Enums

Direction

Defines the direction in which text is to be read.