composable_views

Trait Output

pub trait Output: Sized {
    // Required methods
    fn begin(&mut self, rgba: [u8; 4], transform: &Transform);
    fn move_to(&mut self, x: f32, y: f32);
    fn line_to(&mut self, x: f32, y: f32);
    fn quadratic_bezier_to(&mut self, x1: f32, y1: f32, x: f32, y: f32);
    fn cubic_bezier_to(
        &mut self,
        x1: f32,
        y1: f32,
        x2: f32,
        y2: f32,
        x: f32,
        y: f32,
    );
    fn close(&mut self);
}

A surface, or file format, that views may be rendered to.

Note

If a Dependency::<Transform> is in scope, the Output is expected to transform its input accordingly.

Required Methods

fn begin(&mut self, rgba: [u8; 4], transform: &Transform)

Sets up the uniform parameters of the following paths.

Path should always begin with a move_to.

fn move_to(&mut self, x: f32, y: f32)

Begins a new path.

The path should be continued with a series of line_to, quadratic_bezier_to, and/or cubic_bezier_to calls and ended with a call to close.

fn line_to(&mut self, x: f32, y: f32)

Adds a line to the current path.

fn quadratic_bezier_to(&mut self, x1: f32, y1: f32, x: f32, y: f32)

Adds a quadratic Bézier to the current path.

(x1, y1) represents the Bézier control point.

fn cubic_bezier_to( &mut self, x1: f32, y1: f32, x2: f32, y2: f32, x: f32, y: f32, )

Adds a cubic Bézier to the current path.

(x1, y1) and (x2, y2) represent the Bézier control points.

fn close(&mut self)

Closes the current path.

Once this method has been called there is no current path until begin is called again.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".