composable_views

Trait View

pub trait View: Sized {

    // Required methods
    fn size(&self, within: Size) -> Size;
    fn draw(&self, bounds: Bounds, onto: &mut impl Output);

    // Provided methods
    fn event(&self, event: Event, bounds: Bounds) { ... }
    fn padding(self, top: f32, right: f32, bottom: f32, left: f32) -> impl View { ... }
    fn padding_top(self, pad: f32) -> impl View { ... }
    fn padding_right(self, pad: f32) -> impl View { ... }
    fn padding_bottom(self, pad: f32) -> impl View { ... }
    fn padding_left(self, pad: f32) -> impl View { ... }
    fn padding_horizontal(self, pad: f32) -> impl View { ... }
    fn padding_vertical(self, pad: f32) -> impl View { ... }
    fn padding_both(self, x: f32, y: f32) -> impl View { ... }
    fn padding_all(self, pad: f32) -> impl View { ... }
    fn fixed(self, width: f32, height: f32) -> impl View { ... }
    fn width(self, width: f32) -> impl View { ... }
    fn height(self, height: f32) -> impl View { ... }
    fn opacity<P>(self, opacity: f32) -> Opacity<Self> { ... }
    fn background<P>(self, path: P) -> Background<Self, P> { ... }
    fn adjust_width(&self, width: f32) { ... }
    fn adjust_height(&self, height: f32) { ... }
    fn on_tap<A, E>(self, id: Id, action: A, send: E) -> TapGesture<Self, A, E>
       where A: Clone,
             E: Effects<Action = A> { ... }
    fn on_tap_target<A, E>(
        self,
        id: Id,
        action: A,
        send: E,
        size: Size,
    ) -> Target<TapGesture<Self, A, E>>
       where A: Clone,
             E: Effects<Action = A> { ... }

}

Required Methods

fn size(&self, within: Size) -> Size

The intrinsic size of the View

fn draw(&self, bounds: Bounds, onto: &mut impl Output)

How the View is drawn

Provided Methods

fn event(&self, event: Event, bounds: Bounds)

User-interface Event handling of the View

fn padding(self, top: f32, right: f32, bottom: f32, left: f32) -> impl View

View Layout

Add padding to all sides of the View

fn padding_top(self, pad: f32) -> impl View

View Layout

Add padding to the top of the View

fn padding_right(self, pad: f32) -> impl View

View Layout

Add padding to the right side of the View

fn padding_bottom(self, pad: f32) -> impl View

View Layout

Add padding to the bottom of the View

fn padding_left(self, pad: f32) -> impl View

View Layout

Add padding to the left side of the View

fn padding_horizontal(self, pad: f32) -> impl View

View Layout

Add padding to the horizontal sides of the View

fn padding_vertical(self, pad: f32) -> impl View

View Layout

Add padding to the vertical sides of the View

fn padding_both(self, x: f32, y: f32) -> impl View

View Layout

Add different padding to the horizontal and vertical sides of the View

fn padding_all(self, pad: f32) -> impl View

View Layout

Add the same padding to all sides of the View

fn fixed(self, width: f32, height: f32) -> impl View

View Layout

Set the size of the View to a fixed value.

fn width(self, width: f32) -> impl View

View Layout

Set the width of the View to a fixed value.

fn height(self, height: f32) -> impl View

View Layout

Set the height of the View to a fixed value.

fn opacity<P>(self, opacity: f32) -> Opacity<Self>

View Styling

Draws the View with the given opacity

Note

Opacity values range from zero to one; with zero being fully transparent view and one being fully opaque

fn background<P>(self, path: P) -> Background<Self, P>

View Styling

Add a background shape to the View; defined by a Path

fn adjust_width(&self, width: f32)

Flexible View handling

Flexible views re-compute their size based upon the bounds of their containing view. Spacer::fill for example.

fn adjust_height(&self, height: f32)

Flexible View handling

Flexible views re-compute their size based upon the bounds of their containing view. Spacer::fill for example.

fn on_tap<A, E>(self, id: Id, action: A, send: E) -> TapGesture<Self, A, E>
where A: Clone, E: Effects<Action = A>,

Gesture handling

The View’s response to a tap

fn on_tap_target<A, E>( self, id: Id, action: A, send: E, size: Size, ) -> Target<TapGesture<Self, A, E>>
where A: Clone, E: Effects<Action = A>,

Gesture handling

The View’s response to a tap, but presenting a larger tap surface than is drawn

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".